[ANDROSDK-1964] Move FieldsHelper to network
