+++ core/src/test/java/org/hisp/dhis/android/core/arch/api/fields/internal/FieldsShould.kt
@@ -30,8 +30,8 @@ package org.hisp.dhis.android.core.arch.api.fields.internal
 import com.google.common.truth.Truth
 import nl.jqno.equalsverifier.EqualsVerifier
 import nl.jqno.equalsverifier.Warning
-import org.hisp.dhis.android.core.arch.fields.internal.FieldsHelper
 import org.hisp.dhis.android.network.common.fields.Fields
+import org.hisp.dhis.android.network.common.fields.FieldsHelper
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
